Add package duke, command, gui and task: error encountered
