+++ build.gradle
@@ -49,7 +49,7 @@ repositories {
 
 application {
     // Change this to your main class.
-    mainClassName = "Duke"
+    mainClassName = "duke.Duke"
 }
 
 run {
