+++ data/duke.txt
@@ -2,4 +2,3 @@ E | 1 | project meeting | May 20 2020
 D | 1 | 2103T | May 05 2020
 T | 1 | eat veges
 T | 1 | 2101
-T | 1 | 2102
