+++ src/main/java/duke/command/Done.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.Storage;
+import duke.task.TaskList;
+import duke.Ui;
+
 public class Done extends Command {
 
     private int taskNumber;
