+++ src/main/java/duke/command/Exit.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.Storage;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.Ui;
+
 public class Exit extends Command {
 
     public Exit(Task task) {
