+++ src/main/java/duke/gui/DialogBox.java
@@ -1,3 +1,5 @@
+package duke.gui;
+
 import java.io.IOException;
 import java.util.Collections;
