+++ src/main/java/duke/gui/Launcher.java
@@ -1,3 +1,5 @@
+package duke.gui;
+
 import javafx.application.Application;
 
 /**
@@ -5,6 +7,6 @@ import javafx.application.Application;
  */
 public class Launcher {
     public static void main(String[] args) {
-        Application.launch(Main.class, args);
+        Application.launch(duke.gui.Main.class, args);
     }
 }
