+++ src/test/java/EventTest.java
@@ -1,3 +1,4 @@
+import duke.task.Event;
 import org.junit.jupiter.api.Test;
 
 import java.time.DateTimeException;
