+++ src/test/java/TaskListTest.java
@@ -1,3 +1,8 @@
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.task.ToDo;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
