+++ src/test/java/TaskTest.java
@@ -1,3 +1,4 @@
+import duke.task.Task;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
