+++ src/test/java/ToDoTest.java
@@ -1,3 +1,4 @@
+import duke.task.ToDo;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
