+++ text-ui-test/sources.txt
@@ -1,8 +1,8 @@
-D:\duke\src\main\java\CommandList.java
-D:\duke\src\main\java\Deadline.java
-D:\duke\src\main\java\Duke.java
-D:\duke\src\main\java\DukeErrorType.java
-D:\duke\src\main\java\DukeException.java
-D:\duke\src\main\java\Event.java
-D:\duke\src\main\java\Task.java
-D:\duke\src\main\java\ToDo.java
+D:\duke\src\main\java\duke.command.CommandList.java
+D:\duke\src\main\java\duke.task.Deadline.java
+D:\duke\src\main\java\duke.Duke.java
+D:\duke\src\main\java\duke.DukeErrorType.java
+D:\duke\src\main\java\duke.DukeException.java
+D:\duke\src\main\java\duke.task.Event.java
+D:\duke\src\main\java\duke.task.Task.java
+D:\duke\src\main\java\duke.task.ToDo.java
